// frontend/src/shared/Enums.tsx
export enum ContractTypesEnum {
    PERMANENT = "Permanent",
    CONTRACT = "Contract",
}
export enum WorkTypesEnum {
    FULL_TIME = "Full-time",
    PART_TIME = "Part-time",
}

export enum MonthsEnum {
    JANUARY = "January",
    FEBRUARY = "February",
    MARCH = "March",
    APRIL = "April",
    MAY = "May",
    JUNE = "June",
    JULY = "July",
    AUGUST = "August",
    SEPTEMBER = "September",
    OCTOBER = "October",
    NOVEMBER = "November",
    DECEMBER = "December",
}